refactor: modify directory structure
