+++ client/src/hooks/index.js
@@ -1,0 +1,4 @@
+import useFetch from "./useFetch";
+import useInput from "./useInput";
+
+export { useFetch, useInput };
