+++ client/src/pages/index.js
@@ -1,0 +1,5 @@
+import Home from "./Home";
+import Todo from "./Todo";
+import About from "./About";
+
+export { Home, Todo, About };
